refactor - moved React components into components dir
