+++ client/src/components/Behodler/Swap/TradingBox3/components/ExtendedTextField.tsx
@@ -21,9 +21,9 @@ import {
     Typography,
     Box,
 } from '@material-ui/core'
-import { isNullOrWhiteSpace, formatNumberText, formatSignificantDecimalPlaces } from './jsHelpers'
-import { WalletContext } from '../../../Contexts/WalletStatusContext'
-import API from '../../../../blockchain/ethereumAPI'
+import { isNullOrWhiteSpace, formatNumberText, formatSignificantDecimalPlaces } from '../jsHelpers'
+import { WalletContext } from '../../../../Contexts/WalletStatusContext'
+import API from '../../../../../blockchain/ethereumAPI'
 
 interface DropDownField {
     name: string
